+++ src/tcode_api/schemas/commands/union.py
@@ -13,6 +13,9 @@ from .calibrate_labware_height.latest import (
 from .calibrate_labware_holder.latest import (
     CALIBRATE_LABWARE_HOLDER,
 )
+from .calibrate_labware_well_center.latest import (
+    CALIBRATE_LABWARE_WELL_CENTER,
+)
 from .calibrate_labware_well_depth.latest import (
     CALIBRATE_LABWARE_WELL_DEPTH,
 )
@@ -54,6 +57,7 @@ TCode = Annotated[
     | ADD_PIPETTE_TIP_GROUP
     | ADD_ROBOT
     | ADD_TOOL
+    | CALIBRATE_LABWARE_WELL_CENTER
     | CALIBRATE_LABWARE_WELL_DEPTH
     | CALIBRATE_LABWARE_HEIGHT
     | CALIBRATE_LABWARE_HOLDER